ai-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai>=1.0.0
PyGithub>=2.0.0
httpx>=0.25.0
